refactor(frontend): split SettingsView, BankView, ImportExcelView into sub-components (BL-077)
